Add ping events to detect network outages
If the network between the server and client goes down then it will
look to the client very much like the server simply isn't sending
any events and that nothing is wrong. By adding in a ping value
and an error condition if they're not received we can detect this
failure mode.

Change-Id: Ic302810f04e94ea822a03d9cb20229838363bb9f