Avoid load bug in firefox
Move JS to the bottom so that we can be completely certain that all
the page's elements exist before any of the JS runs. This is needed
since there are some cases where the event stream returns its first
response before the page finishes rendering. That in turn caused the
'tab' variable to be null and the table not to render. Interestingly
the elements were still created and found each time through, they
just never got added to the DOM.

Change-Id: Id4b2f48e7f2e89305ea7ea2bab59a48ca993f8a1